Add new constructors for ioline.
C# does not have a way to implement varargs callback methods through
P/Invoke, so the iol interface needed to be simplified.  Instead of
supporting rich formatting a write method that takes a string and a string
length is sufficient.  The original formatting method is presered (if
necessary).  In the case of the iol_new3 the constructor an implementation
of print_cb_t is provided that forwards to the write_cb_t.  This means
less code churn.